attach.jags {R2jags}	R Documentation

Attach/detach elements of ‘JAGS’ objects to search path

Description

These are wraper functions for attach.bugs and detach.bugs, which attach or detach three-way-simulation array of bugs object to the search path. See attach.all for details.

Usage

  attach.jags(x, overwrite = NA)
  detach.jags()

Arguments

x
An rjags object.

overwrite
If TRUE, objects with identical names in the Workspace (.GlobalEnv) that are masking objects in the database to be attached will be deleted. If NA (the default) and an interactive session is running, a dialog box asks the user whether masking objects should be deleted. In non-interactive mode, behaviour is identical to overwrite=FALSE, i.e. nothing will be deleted.

Details

See attach.bugs for details

Author(s)

Yu-Sung Su suyusung@tsinghua.edu.cn,

References

Sibylle Sturtz and Uwe Ligges and Andrew Gelman. (2005). “R2WinBUGS: A Package for Running WinBUGS from R.” Journal of Statistical Software 3 (12): 1–6.

Examples

  # See the example in ?jags for the usage.

[Package R2jags version 0.7-1 Index]